refactor all the things

+++ netvr-unity/Assets/OtherScripts/IsblNetRemoteClient.cs
@@ -1,8 +1,9 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
 public class IsblNetRemoteClient : MonoBehaviour
 {
-    public readonly Dictionary<int, IsblNetRemoteDevice> Devices = new();
+    public readonly Dictionary<UInt16, IsblNetRemoteDevice> Devices = new();
 }
